chore: move versioning outside main package
