+++ nodebuilder/node_test.go
@@ -13,6 +13,7 @@ import (
 	collectormetricpb "go.opentelemetry.io/proto/otlp/collector/metrics/v1"
 	"google.golang.org/protobuf/proto"
 
+	"github.com/celestiaorg/celestia-node/cmd/version"
 	"github.com/celestiaorg/celestia-node/nodebuilder/node"
 	"github.com/celestiaorg/celestia-node/share"
 )
@@ -81,7 +82,7 @@ func TestLifecycle_WithMetrics(t *testing.T) {
 						otlpmetrichttp.WithInsecure(),
 					},
 					tt.tp,
-					node.BuildInfo{},
+					version.BuildInfo{},
 				),
 			)
 			require.NotNil(t, node)
